Refactor to library
